requirements.txt
Django==3.2
django-cors-headers==3.7.0
django-environ==0.8.1
django-tenant-users==0.4.0
django-tenants==3.4.2
django-timezone-field==4.2.1
djangorestframework==3.12.4
python-dateutil==2.8.2
psycopg2==2.9.5
whitenoise==6.0.0